Remove mobile nav
The separate mobile nav is removed. For mobile devices, The nav within
the header is moved below to acts as the nav for mobile. Hamburger
button to toggle the nav for mobile smaller width devices still works.
The nav when on mobile devices is positioned absolutely so toggling it
won't affect the position of other elements in the page. The lowering of
opacity when mobile nav is not working.